Organize Cary data

Cleans the raw data extracted from the Cary 5000 UV-Vis-IR spectrophotometer, ensuring an organized dataset for analysis. During the process, duplicate columns containing wavelength are removed to avoid redundancy. Samples with missing data are also eliminated, ensuring that only complete measurements are retained. Additionally, the columns are reordered based on the sample names, making it easier to interpret and manipulate the data. The wavelength column is placed as the first column, followed by the sample columns.

Usage

clean_sheet_cary(
  data = data,
  prefix = NULL,
  name_wave = "Wave",
  range_wave = c(380, 2500),
  nm_step = 0.5
)

Arguments

data

A data.frame containing the wavelength and reflectance values. It should be obtained directly from the export of a CSV file.

prefix

Indicates the prefix for columns where there is no sample. It is usually predefined as x or ..

name_wave

Indicates the name of the wavelength column. The default is 'wave'.

range_wave

Is the wavelength range used. The default is from 380 nm to 2500 nm.

nm_step

The increment in nanometers between consecutive wavelength readings. This defines the resolution of the measurements. The default value is 0.5 nm.

Value

Returns an organized data.frame

Examples

# example code
library(OxSR)
data(data_cary)

clean_sheet_cary(data_cary, prefix = "x")
#> # A tibble: 4,241 × 26
#>    wavelength_nm    a1    a2    a3    a4    a5    a6    a7    a8    a9   a10
#>            <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1         2500   37.0  49.2  41.3  65.1  63.1  45.6  46.7  47.2  45.9  49.3
#>  2         2500.  36.0  48.8  40.5  64.6  62.1  45.8  47.1  44.9  45.5  51.7
#>  3         2499   36.0  47.8  42.1  64.8  61.8  46.4  44.7  45.8  45.1  48.8
#>  4         2498.  38.5  49.6  41.3  64.7  62.6  46.2  46.2  44.9  48.2  50.8
#>  5         2498   36.4  48.8  40.5  65.4  62.3  46.4  45.8  44.9  46.4  50.7
#>  6         2498.  35.5  47.4  39.3  64.7  62.8  47.5  45.4  43.9  46.6  49.8
#>  7         2497   37.5  49.1  40.5  66.3  61.4  47.5  47.2  45.5  47.0  52.3
#>  8         2496.  37.5  49.6  41.2  66.2  62.8  46.3  44.0  46.0  46.4  49.4
#>  9         2496   38.0  48.1  40.3  65.1  61.2  47.2  46.3  44.5  44.9  49.9
#> 10         2496.  36.0  48.1  40.4  63.1  60.8  47.4  45.3  45.7  46.4  50.6
#> # ℹ 4,231 more rows
#> # ℹ 15 more variables: a11 <dbl>, a12 <dbl>, a13 <dbl>, a14 <dbl>, a15 <dbl>,
#> #   a16 <dbl>, a17 <dbl>, a18 <dbl>, a19 <dbl>, a20 <dbl>, a21 <dbl>,
#> #   a22 <dbl>, a23 <dbl>, a24 <dbl>, a26 <dbl>

# With pipe

data_cary |> clean_sheet_cary(prefix = "x")
#> # A tibble: 4,241 × 26
#>    wavelength_nm    a1    a2    a3    a4    a5    a6    a7    a8    a9   a10
#>            <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1         2500   37.0  49.2  41.3  65.1  63.1  45.6  46.7  47.2  45.9  49.3
#>  2         2500.  36.0  48.8  40.5  64.6  62.1  45.8  47.1  44.9  45.5  51.7
#>  3         2499   36.0  47.8  42.1  64.8  61.8  46.4  44.7  45.8  45.1  48.8
#>  4         2498.  38.5  49.6  41.3  64.7  62.6  46.2  46.2  44.9  48.2  50.8
#>  5         2498   36.4  48.8  40.5  65.4  62.3  46.4  45.8  44.9  46.4  50.7
#>  6         2498.  35.5  47.4  39.3  64.7  62.8  47.5  45.4  43.9  46.6  49.8
#>  7         2497   37.5  49.1  40.5  66.3  61.4  47.5  47.2  45.5  47.0  52.3
#>  8         2496.  37.5  49.6  41.2  66.2  62.8  46.3  44.0  46.0  46.4  49.4
#>  9         2496   38.0  48.1  40.3  65.1  61.2  47.2  46.3  44.5  44.9  49.9
#> 10         2496.  36.0  48.1  40.4  63.1  60.8  47.4  45.3  45.7  46.4  50.6
#> # ℹ 4,231 more rows
#> # ℹ 15 more variables: a11 <dbl>, a12 <dbl>, a13 <dbl>, a14 <dbl>, a15 <dbl>,
#> #   a16 <dbl>, a17 <dbl>, a18 <dbl>, a19 <dbl>, a20 <dbl>, a21 <dbl>,
#> #   a22 <dbl>, a23 <dbl>, a24 <dbl>, a26 <dbl>